requirements.txt
torch
numpy
trimesh
pyglet<2
scipy
matplotlib
pyopengl
pyqt6